test(doctor): reuse the legacy recovery test root
